combine market_case_1.ts/market_case_1.test.ts -> nft_market.test.ts
also put some important code into beforeEach block
